tests/wpt/web-platform-tests/css/motion-1/parsing/offset-path-parsing-invalid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>Motion Path Module Level 1: parsing offset-path with invalid values</title>
<link rel="author" title="Eric Willigers" href="mailto:ericwilligers@chromium.org">
<link rel="help" href="https://drafts.fxtf.org/motion-1/#offset-path-property">
<meta name="assert" content="offset-path supports only the grammar from the spec.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/parsing-testcommon.js"></script>
</head>
<body>
<script>
// arc path segments must have at least 7 arguments.
// https://www.w3.org/TR/SVG/paths.html#PathDataEllipticalArcCommands
test_invalid_value("offset-path", "path('M 20 30 A 60 70 80')");

test_invalid_value("offset-path", "ray(0 sides)");
test_invalid_value("offset-path", "ray(0deg)");
test_invalid_value("offset-path", "ray(closest-side)");
test_invalid_value("offset-path", "ray(closest-side 0deg closest-side)");
test_invalid_value("offset-path", "ray(0deg closest-side 0deg)");
test_invalid_value("offset-path", "ray(contain 0deg closest-side contain)");

</script>
</body>
</html>
